requirement.txt
numpy
torch
matplotlib
Pillow
tkinter